Debian packaging for Stats 2.0.0
This commit represents the final Debian packaging for Stats 2.0.0. It can be
used to build a package for Debian, Ubuntu, or any other Debian derivative.
Since the script is written in Perl and the package is arch all, the resulting
package will work on virtually any Debian derivative without recompilation.