Added an additional screenshot utility.
Not only did I add another screenshot utility to print_screenshot(), but I
made it the default. scrot is now the preferred screeshot utility, with
import relegated to a backup option if scrot is not installed.

The ImageMagick import utility works fairly well, but it sometimes results in
black areas of a screenshot when a compositor is in use. Compiz
(used by Unity) and Mutter (the GNOME Shell compositor) don't seem to perturb
it, but Compton (which I use with MATE) and Gala (the Elementary OS/Pantheon
compositor) cause the problems described above, which basically ruins any
screenshot.

I originally used import to take screenshots to maintain compatability with
Stats 1.x. However now that Stats 2.0 has seen some real-world usage it
appears that this bug is too widespread to ignore. Since scrot is another
screenshot utility with a fairly large install base - and, importantly, does
not suffer from this problem - I decided to make it default.