Added --output-dir option.
This commit introduces a new option to control the output directory for
screenshots and reports. Prior to the addition of this switch the user could
only control the output directory for the files this script generates
implicitly by changing to the desired output directory before executing the
script. The --output-dir switch makes it possible for the user specify the
output directory explicitly without using the aforementioned workaround.

The man page and builtin help information have also both been updated to
document the new switch.