Added Elementary OS detection and logo.
Note: The PNG version of the Elementary OS logo in the logos/ directory is not
quite the same as the original. I generated it from the official SVG, but I
made the borders thicker. No software I tried could properly convert the
original version of the logo to ascii - its was apparently just too thin!