Added a space before the stats printout when --no-logo is in effect.
The statistics and screenshot functions start printing data without first
printing a space. They assumed that if a space was supposed to be printed
before them, it was done outside of the function. This results in everything
being spaced properly (one space between each section) when the distro's logo
is printed because each logo has one space before and after it for balanace.
However when the --no-logo switch was passed to the script, this consistency
was broken because the logo printing function was exclusively responsible for
printing the initial space. As of this commit if no logo is printed spacing
will still be correct.