Implemented proper support for Elementary OS's Pantheon Desktop Environment.
Although I added the Elementary OS logo and os_release detection to Stats 2.1,
I did not account for the first-party desktop environment it ships with. Since
Elementary OS's desktop environment - Pantheon - extensively uses components
of GNOME, the script was reporting the DE as GNOME.

Thanks to the bug report and debug log supplied by keuka and the initial patch
written by Haggis, Pantheon is now detected reliably. Its version number is
also detected as long Pantheon is running on Ubuntu (such as in the default
Elementary OS Luna configuration), but since version detection is tied to APT,
it will not be detected properly on non-Debian-based distributions (such as
Arch Linux). This limitation will hopefully be lifted when Pantheon gets proper
version information in the binary.

I am committing this patch on behalf of Haggis, but the content of this
commit message is my own.